requirements.txt
pandas
numpy

streamlit
#streamlit-drawable-canvas

torch
transformers
huggingface_hub[hf_xet]

pillow
